Создает taskController, переносит в него функцию рендера таска. Подключает taskController в boardController
